forever working on syntax

+++ js/break_and_continue.js
@@ -1,7 +1,7 @@
  "use strict"
 let chosen = null
 while (chosen === null) {
-     chosen = parseFloat(prompt('please input an odd number between 1 and 50'))
+     chosen = parseInt(prompt('please input an odd number between 1 and 50'))
     if (isNaN(chosen) == true) chosen = null
     else if (chosen < 0 || chosen > 50) chosen = null
     else if (chosen % 2 !== 0) break;
